minor refactor: avoid exporting in +Page.tsx
