+++ docs/pages/index/primaryColor.tsx
@@ -1,0 +1,1 @@
+export const primaryColor = '#1810FF'
