+++ docs/pages/index/sections/Flexible/Flexible.tsx
@@ -1,6 +1,7 @@
 import React from 'react'
 import { TextBox } from '../../components/TextBox'
-import { Grid, primaryColor } from '../../+Page'
+import { Grid } from '../../Grid'
+import { primaryColor } from '../../primaryColor'
 import { SectionTextCollection } from '../../components/SectionTextCollection'
 import { ParagraphTextCollection } from '../../components/ParagraphTextCollection'
 import SlotMachineSVG from '../../components/SlotMachineSVG'
